isolate utility functions
